Add files via upload
The Automatic Water Dispenser project is a smart and innovative solution designed to automate the dispensing of water, promoting convenience and efficiency in water usage. This project leverages modern technologies to ensure accurate measurement and user-friendly operation, making it ideal for home, office, or public space use.The IoT Based Automatic Water Dispenser project harnesses the power of the Internet of Things (IoT) to provide a smart and efficient solution for managing water dispensing. This project allows users to remotely control and monitor water usage through a mobile app or web interface, making it perfect for homes, offices, and public areas.